refactor into linux specific implementation
Define high level system interface that allows us to provide multiple
implementaitons for the disko API.

Refactor the existing disk implementation into a linux specific
implementation.
