+++ disk_test.go
